requirements.txt
beautifulsoup4==4.9.3
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
opencv-contrib-python==4.5.1.48
Pillow==8.1.2
pkg-resources==0.0.0
soupsieve==2.2.1
Werkzeug==1.0.1
